+++ Physics/Source/WeaponScript.cpp
@@ -91,7 +91,7 @@ void WeaponScript::UpdateStats(GameObject* go, bool Multiply)
 		}
 		case PartScript::CLIP:
 		{
-			m_iMagazineRounds_Max = m_iMagazineRounds_Max + go->PART->GetMultiplier();
+			m_iMagazineRounds_Max = m_iMagazineRounds_Max + static_cast<int>(go->PART->GetMultiplier());
 			break;
 		}
 		case PartScript::GRIP:
@@ -120,7 +120,7 @@ void WeaponScript::UpdateStats(GameObject* go, bool Multiply)
 		}
 		case PartScript::CLIP:
 		{
-			m_iMagazineRounds_Max = m_iMagazineRounds_Max - go->PART->GetMultiplier();
+			m_iMagazineRounds_Max = m_iMagazineRounds_Max - static_cast<int>(go->PART->GetMultiplier());
 			Math::Clamp(m_iMagazineRounds, m_iMagazineRounds, m_iMagazineRounds_Max);
 
 			break;
@@ -175,26 +175,26 @@ void WeaponScript::AddPart(GameObject* part)
 		case PartScript::SCOPE:
 		{
 			m_ScopeParts.push_back(part);
-			part->TRANS->SetRelativePosition(0.5 * m_ScopeParts.size(), 0.5, 0);
+			part->TRANS->SetRelativePosition(0.5f * m_ScopeParts.size(), 0.5f, 0.f);
 			break;
 		}
 		case PartScript::MUZZLE:
 		{
 			m_MuzzleParts.push_back(part);
-			part->TRANS->SetRelativePosition(1 * m_MuzzleParts.size(), 0, 0);
+			part->TRANS->SetRelativePosition(1.f * m_MuzzleParts.size(), 0, 0);
 
 			break;
 		}
 		case PartScript::CLIP:
 		{
 			m_StockParts.push_back(part);
-			part->TRANS->SetRelativePosition(0, -0.5 * m_StockParts.size(), 0);
+			part->TRANS->SetRelativePosition(0, -0.5f* m_StockParts.size(), 0);
 			break;
 		}
 		case PartScript::GRIP:
 		{
 			m_GripParts.push_back(part);
-			part->TRANS->SetRelativePosition((-1.0 * m_GripParts.size()), 0, 0);
+			part->TRANS->SetRelativePosition((-1.0f * m_GripParts.size()), 0, 0);
 			break;
 		}
 		default:
